requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-shortuuidfield==0.1.3
djangorestframework==3.10.3
idna==2.8
mysqlclient==1.4.4
Pillow==6.2.0
python3-memcached==1.51
pytz==2019.2
qiniu==7.2.4
requests==2.22.0
shortuuid==0.5.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6

